server/requirements.txt
# server
Pillow~=11.2.0
black~=25.1.0
playwright~=1.52.0
python-dotenv~=0.9.0
